Use long long into to match original int64_t type

+++ include/fields2cover/route_planning/route_planner_base.h
@@ -65,12 +65,12 @@ class RoutePlannerBase {
 
   /// Use the optimizer to generate the index of the points of the best
   ///   coverage route.
-  virtual std::vector<long int> computeBestRoute(
+  virtual std::vector<long long int> computeBestRoute(
       const F2CGraph2D& cov_graph, bool show_log, long int time_limit_seconds) const;
 
   /// Tranform index of points to an actual Route.
   virtual F2CRoute transformSolutionToRoute(
-      const std::vector<long int>& route_ids,
+      const std::vector<long long int>& route_ids,
       const F2CSwathsByCells& swaths_by_cells,
       const F2CGraph2D& coverage_graph,
       F2CGraph2D& shortest_graph) const;
